pyproject.toml
[tool.poetry]
name = "harmonica-l12-node"
version = "0.1.0"
description = ""
authors = ["aoirint <aoirint@gmail.com>"]
readme = "README.md"
packages = [{include = "harmonica_l12_node"}]

[tool.poetry.dependencies]
python = "~3.11"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
html5lib = "^1.1"
schedule = "^1.2.0"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
